+++ src/layout/AppMobileLayout.tsx
@@ -1,5 +1,6 @@
 import { type MobileNavBarTitleProps } from '@lobehub/ui';
 import { createStyles } from 'antd-style';
+import Head from 'next/head';
 import { CSSProperties, PropsWithChildren, ReactNode, memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
@@ -45,21 +46,29 @@ const AppMobileLayout = memo<AppMobileLayoutProps>(
     const { styles, cx } = useStyles();
 
     return (
-      <Flexbox className={cx(styles.container, className)} style={style}>
-        {navBar && (
-          <>
-            <div className={styles.mobileNavBar}>{navBar}</div>
-            <SafeSpacing mobile position={'top'} />
-          </>
-        )}
-        {children}
-        {showTabBar && (
-          <>
-            <SafeSpacing mobile position={'bottom'} />
-            <MobileTabBar className={styles.mobileTabBar} />
-          </>
-        )}
-      </Flexbox>
+      <>
+        <Head>
+          <meta
+            content="width=device-width, initial-scale=1.0, viewport-fit=cover"
+            name="viewport"
+          />
+        </Head>
+        <Flexbox className={cx(styles.container, className)} style={style}>
+          {navBar && (
+            <>
+              <div className={styles.mobileNavBar}>{navBar}</div>
+              <SafeSpacing mobile position={'top'} />
+            </>
+          )}
+          {children}
+          {showTabBar && (
+            <>
+              <SafeSpacing mobile position={'bottom'} />
+              <MobileTabBar className={styles.mobileTabBar} />
+            </>
+          )}
+        </Flexbox>
+      </>
     );
   },
 );
